chore(front): move mocked-metadata helpers under src/testing
After #20308 production code no longer loads mocked metadata when the
user is unauthenticated. The remaining mocked-metadata helpers are only
consumed by Storybook decorators, so move them next to the other
testing-only utilities to make their scope explicit and prevent
accidental re-introduction of a runtime dependency on mocks.

- src/modules/metadata-store/hooks/useLoadMockedMetadata.ts
  -> src/testing/hooks/useLoadMockedMetadata.ts
- src/modules/metadata-store/utils/preloadMockedMetadata.ts
  -> src/testing/utils/preloadMockedMetadata.ts

Update the three Storybook decorator imports accordingly. No runtime
behavior change.
